updated yahoo.py and added functions
The yahoo.py file was changed to add the Stock_Info, because I want to be able to save this data to a csv so that I can use it for the dashboard.

An interesting problem came up that I didn't account for. If this code tries to run on a weekend it won't work because the stock market is closed. I added some functions to determine what day of the week it is. If the week day number is 6 or 0 it is the weekend so I need to make the code not run for those days